Code refactoring WIP.
