+++ src/Controller/UserController.php
@@ -10,11 +10,8 @@ use App\Service\RegistrationManager;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
 use Symfony\Component\Translation\TranslatorInterface;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class UserController extends Controller
 {
@@ -104,70 +101,52 @@ class UserController extends Controller
             return $this->render('user/notFound.html.twig');
         }
 
-        //\var_dump($user);
-
-        //$avatar_image_file = $user->getId() . "." . $user->getFileExtension();
-
         $form = $this->createForm(UserType::class, $user);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            // $file stores the uploaded image file
-            /** @var Symfony\Component\HttpFoundation\File\UploadedFile $file */
-            $file = $user->getAvatar();
+            $destPage = self::handleUserSubmit($user);
 
-            if ($file !== null) {
-                $fileName = $user->getId().'.'.$file->guessExtension();
-
-                $user->setFileExtension($file->guessExtension());
-
-                // Move the file to the directory where brochures are stored
-                try {
-                    $file->move(
-                        $this->getParameter('avatars_directory'),
-                        $fileName
-                    );
-                } catch (FileException $e) {
-                    // ... handle exception if sth happens during file upload
-                }
-            }
-
-            // updates the 'avatar' property to store the image file name
-            // instead of its contents
-            //$user->setAvatar($fileName);
-
-            $result = $this->userManager->saveUserToDB($user);
-
-            $request->getSession()->getFlashBag()->add(
-                $result['msg_type'],
-                $this->i18n->trans($result['message'], [], 'gui')
-            );
-
-            return $this->redirectToRoute($result['dest_page']);
+            return $this->redirectToRoute($destPage);
         }
 
         return $this->render('user/edit.html.twig', [
             'form' => $form->createView(),
-            'user' => $user,
-            //'avatar_image' => $avatar_image_file,
+            'user' => $user
         ]);
     }
 
-    /**
-     * @Route("/login", name="user_login")
-     */
-    public function login(AuthenticationUtils $authenticationUtils): Response
+    private function handleUserSubmit(User $user)
     {
-        $data = $this->userManager->initLoginForm($authenticationUtils);
+        // $file stores the uploaded image file
+        /** @var Symfony\Component\HttpFoundation\File\UploadedFile $file */
+        $file = $user->getAvatar();
 
-        return $this->render(
-            'user/login.html.twig',
-            [
-                'last_username' => $data['lastUsername'],
-                'error' => $data['error'],
-            ]
+        if ($file !== null) {
+            $fileName = $user->getId().'.'.$file->guessExtension();
+
+            $user->setFileExtension($file->guessExtension());
+
+            // Move the file to the directory where brochures are stored
+            try {
+                $file->move(
+                    $this->getParameter('avatars_directory'),
+                    $fileName
+                );
+            } catch (FileException $e) {
+                // ... handle exception if sth happens during file upload
+            }
+        }
+
+        $result = $this->userManager->saveUserToDB($user);
+
+        $request->getSession()->getFlashBag()->add(
+            $result['msg_type'],
+            $this->i18n->trans($result['message'], [], 'gui')
         );
+        
+        return $result['dest_page'];
     }
 
     /**
@@ -230,12 +209,7 @@ class UserController extends Controller
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            if (true === $this->pwdManager->checkAndSaveNewPwd(
-                $request->request->get('m'),
-                $request->request->get('t'),
-                $request
-            )
-            ) {
+            if (true === $this->pwdManager->checkAndSaveNewPwd($request)) {
                 $request->getSession()->getFlashBag()->add(
                     'success',
                     $this->i18n->trans('new_pwd_successfuly_saved')
@@ -276,26 +250,5 @@ class UserController extends Controller
         );
 
         return $this->redirectToRoute('user_forgotten_pwd');
-    }
-    
-    /**
-     * @Route("/logout", name="app_logout", methods={"GET"})
-     */
-    public function logout()
-    {
-        // controller can be blank: it will never be executed!
-        throw new \Exception('Don\'t forget to activate logout in security.yaml');
-    }
-
-    // Without this method, a 404 error occurs before the firewall is called.
-    // The result is that a connected user is not detected anymore on the
-    // 404 error page, and he appears disconnected.
-    // See https://github.com/symfony/symfony/issues/5320#issuecomment-56401080
-    /**
-     * @Route("/{something}", name="page_not_found")
-     */
-    public function pageNotFound()
-    {
-        throw new NotFoundHttpException();
     }
 }
