+++ templates/security/login.html.twig
@@ -1,4 +1,4 @@
-{# templates/user/login.html.twig #}
+{# templates/security/login.html.twig #}
 {% trans_default_domain 'gui' %}
 {% extends 'layout.html.twig' %}
 {% block content %}
